ci: fix the four red jobs + lint cleanups
- ci: bump golangci-lint-action v6 -> v7 (v6 cannot drive golangci-lint v2)
- goreleaser: rename `checksums:` -> `checksum:` (GoReleaser v2.10 schema)
- install.sh: use $HOME not ~ inside quotes (SC2088); disable the intentional SC2016
- tests: bind unix sockets under /tmp via a shortSock helper so macOS's 104-byte
  sun_path limit isn't exceeded under a long $TMPDIR (the macOS CI runner) — fixes
  the build-test-lint (macos-latest) `go test -race` failures
- lint: golangci-lint --fix (gofmt, misspell, unconvert, staticcheck quickfixes);
  check the stderr Fprintf errors (errcheck); allow descriptive exported names
  (revive disableStutteringCheck — ingest.IngestItem would otherwise be flagged)

All gates green locally: build, vet, staticcheck, golangci-lint (0 issues),
test -race (11/11), shellcheck, goreleaser check.

+++ internal/httpd/socket_unix.go
@@ -37,7 +37,7 @@ func listenUnixSocket(path string) (net.Listener, error) {
 		if !ok {
 			return nil, fmt.Errorf("clipbeam: cannot stat existing socket %q", path)
 		}
-		if uint32(st.Uid) != uint32(os.Getuid()) {
+		if st.Uid != uint32(os.Getuid()) {
 			return nil, fmt.Errorf("clipbeam: refusing to bind socket %q owned by uid %d (not %d)", path, st.Uid, os.Getuid())
 		}
 		if err := os.Remove(path); err != nil {
